first pass refactor to data model structure
