+++ views/partials/header.html
@@ -1,4 +1,4 @@
 <div class="header js-header" id="header">
-    <div class="typography--head">{{variables.heading}}</div>
-    <div class="typography--large primary-color">{{variables.sub_heading}}</div>
+    <div class="typography--head">{{data.variables.heading}}</div>
+    <div class="typography--large primary-color">{{data.variables.sub_heading}}</div>
 </div>
